day12/requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
pypdf
python-dotenv
sentence-transformers
